# API_User/dist/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User / Avis API",
    "version": "1.0.0",
    "description": "Documentation de l'API de NaoLibre"
  },
  "servers": [
    {
      "url": "http://localhost:8081/naolibre"
    }
  ],
...
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Recuperer tous les utilisateurs",
        "description": "Endpoint pour recuperer tous les utilisateurs",
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Ajouter un utilisateur",
        "description": "Endpoint pour ajouter un utilisateur",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "Informations de l'utilisateur",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "example": {
                          "type": "string",
                          "example": "test"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "example": {
                          "type": "string",
                          "example": "example@example.example"
                        }
                      }
                    },
                    "password": {
                      "type": "object",
                      "properties": {
                        "example": {
                          "type": "string",
                          "example": "example"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Utilisateur ajouté avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Utilisateur ajouté avec succès"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur lancée par l'API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "L'utilisateur existe déjà /OU/ Email or password is null /OU/ Autre erreur"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "example"
                  },
                  "email": {
                    "type": "string",
                    "example": "example@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "example"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/avis/{nomLigne}": {
      "get": {
        "tags": [
          "Avis"
        ],
        "summary": "Recuperer un avis",
        "description": "Endpoint pour recuperer un avis par le nom de la ligne",
        "parameters": [
          {
            "name": "nomLigne",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nom de la ligne"
          }
        ],
        "responses": {
          "200": {
            "description": "Avis trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LigneModel"
                }
              }
            }
          },
          "401": {
            "description": "Cette ligne n'a pas d'avis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Cette ligne n'a pas d'avis"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/avis/{nomLigne}/{email}/{password}": {
      "post": {
        "tags": [
          "Avis"
        ],
        "summary": "Ajouter un avis",
        "description": "Endpoint pour ajouter un avis",
        "parameters": [
          {
            "name": "nomLigne",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nom de la ligne"
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Avis",
            "in": "body",
            "description": "Informations de l'avis",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "note": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "example": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Avis ajouté avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvisModel"
                }
              }
            }
          },
          "400": {
            "description": "Erreur lancer par l'API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "description": "Avis is null /OU/ Ligne or email or password are null /OU/ email or password are not correct /OU/ the note is not correct /OU/ Error"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserModel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              }
            }
          },
          "email": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "example": {
                "type": "string",
                "example": "example@example.example"
              }
            }
          },
          "password": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "nullable": {
                "type": "boolean",
                "example": true
              },
              "example": {
                "type": "string",
                "example": "password"
              }
            }
          },
          "point": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "integer"
              },
              "example": {
                "type": "number",
                "example": 0
              }
            }
          },
          "favori": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "items": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Cardo"
                  }
                }
              }
            }
          }
        }
      },
...
      "LigneModel": {
        "type": "object",
        "properties": {
          "nomLigne": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "string"
              },
              "example": {
                "type": "string",
                "example": "c6"
              }
            }
          },
          "avis": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "array"
              },
              "items": {
                "$ref": "#/components/schemas/AvisModel"
              }
            }
          }
        }
      }
    }
  }
}